+++ src/mako/redis/ConnectionManager.php
@@ -36,6 +36,15 @@ class ConnectionManager extends BaseConnectionManager
 
 		$config = $this->configurations[$connection];
 
-		return new Redis(new Connection($config['host'], $config['port'], $config['persistent'] ?? false, $config['timeout'] ?? 60, $config['connection_timeout'] ?? 5, $connection), $config);
+		$options =
+		[
+			'name'               => $connection,
+			'persistent'         => $config['persistent'] ?? false,
+			'connection_timeout' => $config['connection_timeout'] ?? 5,
+			'read_write_timeout' => $config['timeout'] ?? 60,
+			'nodelay'            => $config['nodelay'] ?? true,
+		];
+
+		return new Redis(new Connection($config['host'], $config['port'], $options), $config);
 	}
 }