+++ src/mako/error/handlers/web/ProductionHandler.php
@@ -144,7 +144,7 @@ class ProductionHandler implements HandlerInterface
 			}
 		}
 
-		return $this->view->render('mako-error::' . $view);
+		return $this->view->clearAutoAssignVariables()->render('mako-error::' . $view);
 	}
 
 	/**